V9

+++ sources/ocf_topics.php
@@ -41,7 +41,7 @@ function ocf_get_topic_where($topic_id)
  * @param  ?MEMBER	The member (NULL: current member).
  * @return boolean	The answer.
  */
-function ocf_may_make_personal_topic($member_id=NULL)
+function ocf_may_make_private_topic($member_id=NULL)
 {
 	if (is_null($member_id)) $member_id=get_member();
 
@@ -164,7 +164,8 @@ function ocf_has_read_topic($topic_id,$topic_last_time=NULL,$member_id=NULL,$mem
 	// Occasionally we need to delete old entries
 	if (mt_rand(0,1000)==123)
 	{
-		$GLOBALS['FORUM_DB']->query('DELETE FROM '.$GLOBALS['FORUM_DB']->get_table_prefix().'f_read_logs WHERE l_time<'.strval($seven_days_ago));
+		if (!$GLOBALS['SITE_DB']->table_is_locked('f_read_logs'))
+			$GLOBALS['FORUM_DB']->query('DELETE FROM '.$GLOBALS['FORUM_DB']->get_table_prefix().'f_read_logs WHERE l_time<'.strval($seven_days_ago));
 	}
 
 	if ($topic_last_time<$seven_days_ago) return true; // We don't store that old
